修正了AI2系列的ScanChessBorad的逻辑漏洞

+++ buweiqi/AI3_Header/MCTS.h
@@ -152,7 +152,7 @@ namespace MCTS
 	Node<State>::~Node()
 	{
 		if (children.size() != 0) {
-			int temp = children.size();
+			int temp = static_cast<int>(children.size());
 			for (int i = 0; i < temp; ++i) {
 				delete children[i];
 			}
